test(sparreal-kernel): move test_hex_fmt to tests dir
sparreal-os/crates/sparreal-kernel ✗ cargo test --target x86_64-unknown-linux-gnu
memory allocation of 24 bytes failed
error: test failed, to rerun pass `--lib`

Caused by:
  process didn't exit successfully: `/rust/tmp/os-checker/repos/sparreal-os/target/x86_64-unknown-linux-gnu/debug/deps/sparreal_kernel-6bac1843bfd664f2` (s
ignal: 6, SIGABRT: process abort signal)
